refactor: move all intefaces into one folder
