+++ src/app/components/user-results-table/user-results-table.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, ViewChild, AfterViewInit, Input } from '@angular/cor
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
-import { Test } from 'src/app/interfaces/test';
+import { Test } from '../../../interfaces/test';
 import { Level } from '../../../constants/data-constants';
 
 @Component({
