+++ src/app/models/common-test/common-test.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { TestStoreService } from '../../services/test/test/test-store.service';
-
-import { Test } from '../../interfaces/test';
+import { map } from 'rxjs/operators';
+import { TestStoreService } from '../../services/store/test-store.service';
+import { Test } from '../../../interfaces/test';
 
 @Component({
   selector: 'app-common-test',
@@ -12,12 +12,17 @@ import { Test } from '../../interfaces/test';
 export class CommonTestComponent implements OnInit {
   test$!: Observable<Test | null>;
 
+  testGrammar$!: Partial<Observable<Test | null>>;
+
+  testListening$!: Partial<Observable<Test | null>>;
+
   selectedIndex = 0;
 
   constructor(private testStoreService: TestStoreService) {}
 
   ngOnInit() {
-    // this.test$ = this.testStoreService.test$;
+    this.test$ = this.testStoreService.test$;
+    // this.testGrammar$ = this.test$.pipe(map((test) => test.grammarQuestions));
   }
 
   setTabIndex(ind: number): void {
