+++ src/app/models/listening-test/listening-test.component.ts
@@ -10,7 +10,7 @@ import {
 import { Track } from 'ngx-audio-player';
 import { Observable } from 'rxjs';
 import { repeat } from 'rxjs/operators';
-import { Test } from '../../interfaces/test';
+import { Test } from '../../../interfaces/test';
 import { MOCK_QUESTION_LIST_AUDITION, Question } from '../../../constants/mock-grammar-test';
 
 @Component({
