+++ src/app/models/speaking-test/speaking-test.component.ts
@@ -4,7 +4,7 @@ import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { ErrorStoreService } from '../../services/store/error-store.service';
 import { Route } from '../../../constants/route-constant';
-import { Test } from '../../interfaces/test';
+import { Test } from '../../../interfaces/test';
 
 declare let MediaRecorder: any;
