+++ src/app/services/auth-http.service.ts
@@ -1,8 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { Injectable } from '@angular/core';
-import { UserResponseType } from '../../interfaces/user.interfaces';
-import { SignIn } from '../interfaces/user';
+import { SignIn, UserResponseType } from '../../interfaces/user.interfaces';
 import { LoginApiUrl, RefreshUserDataUrl } from '../../constants/route-constant';
 
 @Injectable({
