+++ src/interfaces/error.ts
